BufferedReader was the wrong abstraction && use recv1 for initial exchange
BufferedReader is for a normal reader that you want to be buffered.
What's going on here is something else, where the kernel will only send
discrete messages. Dealing with that case directly also allows us to not
copy data around and instead just receive it from the kernel once and
deal with it in-place thereafter.